+++ ApptentiveConnect/source/ATConnect.h
@@ -13,7 +13,7 @@
 #import <Cocoa/Cocoa.h>
 #endif
 
-#define kATConnectVersionString @"1.5.0"
+#define kATConnectVersionString @"1.5.7"
 
 #if TARGET_OS_IPHONE
 #	define kATConnectPlatformString @"iOS"
@@ -28,15 +28,23 @@ extern NSString *const ATMessageCenterUnreadCountChangedNotification;
 /** Notification sent when the user has agreed to rate the application. */
 extern NSString *const ATAppRatingFlowUserAgreedToRateAppNotification;
 
+/** Notification sent when a survey is shown. */
+extern NSString *const ATSurveyShownNotification;
+
 /** Notification sent when a survey is submitted by the user. */
-/** The userInfo dictionary will have a key named `ATSurveyIDKey`, with a value of the id of the survey that was sent. */
 extern NSString *const ATSurveySentNotification;
+
+/**
+ When a survey is shown or sent, notification's userInfo dictionary will contain the ATSurveyIDKey key.
+ Value is the ID of the survey that was shown or sent.
+ */
 extern NSString *const ATSurveyIDKey;
 
 /** Keys for supported 3rd-party integrations. */
 extern NSString *const ATIntegrationKeyUrbanAirship;
 extern NSString *const ATIntegrationKeyKahuna;
 extern NSString *const ATIntegrationKeyAmazonSNS;
+extern NSString *const ATIntegrationKeyParse;
 
 /**
  `ATConnect` is a singleton which is used as the main point of entry for the Apptentive service.
@@ -83,6 +91,7 @@ extern NSString *const ATIntegrationKeyAmazonSNS;
  * `ATIntegrationKeyUrbanAirship` - For Urban Airship
  * `ATIntegrationKeyAmazonSNS` - For Amazon SNS
  * `ATIntegrationKeyKahuna` - For Kahuna
+ * `ATIntegrationKeyParse` - For Parse
  */
 @interface ATConnect : NSObject {
 @private
@@ -95,7 +104,6 @@ extern NSString *const ATIntegrationKeyAmazonSNS;
 	NSMutableDictionary *customDeviceData;
 	NSMutableDictionary *integrationConfiguration;
 	NSString *apiKey;
-	BOOL showTagline;
 	BOOL showEmailField;
 	NSString *initialUserName;
 	NSString *initialUserEmailAddress;
@@ -127,8 +135,6 @@ extern NSString *const ATIntegrationKeyAmazonSNS;
 ///---------------------------------
 /// @name Interface Customization
 ///---------------------------------
-/** Toggles much of the Apptentive branding on and off. `YES` by default. */
-@property (nonatomic, assign) BOOL showTagline;
 /** Toggles the display of an email field in the message panel. `YES` by default. */
 @property (nonatomic, assign) BOOL showEmailField;
 /** Set this if you want some custom text to appear as a placeholder in the feedback text box. */
@@ -145,6 +151,12 @@ extern NSString *const ATIntegrationKeyAmazonSNS;
  @note This setting will be overridden by server-based configuration when it is downloaded.
  */
 @property (nonatomic, assign) BOOL initiallyUseMessageCenter;
+/**
+ Set this to NO to hide Apptentive branding locally on the first launch of your app.
+ 
+ @note This setting will be overridden by server-based configuration when it is downloaded.
+ */
+@property (nonatomic, assign) BOOL initiallyHideBranding;
 #if TARGET_OS_IPHONE
 /**
  A tint color to use in Apptentive-specific UI.
@@ -208,6 +220,25 @@ extern NSString *const ATIntegrationKeyAmazonSNS;
 - (BOOL)engage:(NSString *)eventLabel fromViewController:(UIViewController *)viewController;
 
 /**
+ Engages an event along with custom data about that event. Interaction UI may be shown, if applicable, for the event.
+ 
+ @param eventLabel A string representing the name of the event.
+ @param customData A dictionary of key/value pairs to be associated with the event. Keys and values should conform to standards of NSJSONSerialization's `isValidJSONObject:`.
+ @param viewController A view controller Apptentive UI may be presented from.
+ */
+- (BOOL)engage:(NSString *)eventLabel withCustomData:(NSDictionary *)customData fromViewController:(UIViewController *)viewController;
+
+/**
+ Engages an event along with custom data and extended data about that event. Interaction UI may be shown, if applicable, for the event.
+ 
+ @param eventLabel A string representing the name of the event.
+ @param customData A dictionary of key/value pairs to be associated with the event. Keys and values should conform to standards of NSJSONSerialization's `isValidJSONObject:`.
+ @param extendedData An array of dictionaries with specific Apptentive formatting. For example, [ATConnect extendedDataDate:[NSDate date]].
+ @param viewController A view controller Apptentive UI may be presented from.
+ */
+- (BOOL)engage:(NSString *)eventLabel withCustomData:(NSDictionary *)customData withExtendedData:(NSArray *)extendedData fromViewController:(UIViewController *)viewController;
+
+/**
  Dismisses the message center. You normally won't need to call this.
  
  @param animated `YES` to animate the dismissal, otherwise `NO`.
@@ -227,6 +258,61 @@ extern NSString *const ATIntegrationKeyAmazonSNS;
  */
 - (IBAction)showFeedbackWindow:(id)sender;
 #endif
+
+///--------------------
+/// @name Extended Data for Events
+///--------------------
+
+/**
+ Extended data dictionary representing a point in time, to be included in an event's extended data.
+ 
+ @param date A date and time to be included in an event's extended data.
+ */
++ (NSDictionary *)extendedDataDate:(NSDate *)date;
+
+/**
+ Extended data dictionary representing a location, to be included in an event's extended data.
+ 
+ @param latitude A location's latitude coordinate.
+ @param longitude A location's longitude coordinate.
+ */
++ (NSDictionary *)extendedDataLocationForLatitude:(double)latitude longitude:(double)longitude;
+
+/**
+ Extended data dictionary representing a commerce transaction, to be included in an event's extended data.
+ 
+ @param transactionID The transaction's ID.
+ @param affiliation The store or affiliation from which this transaction occurred.
+ @param revenue The transaction's revenue.
+ @param shipping The transaction's shipping cost.
+ @param tax Tax on the transaction.
+ @param currency Currency for revenue/shipping/tax values.
+ @param commerceItems An array of commerce items contained in the transaction. Create commerce items with [ATConnect extendedDataCommerceItem...].
+ */
++ (NSDictionary *)extendedDataCommerceWithTransactionID:(NSString *)transactionID
+											affiliation:(NSString *)affiliation
+												revenue:(NSNumber *)revenue
+											   shipping:(NSNumber *)shipping
+													tax:(NSNumber *)tax
+											   currency:(NSString *)currency
+										  commerceItems:(NSArray *)commerceItems;
+
+/**
+ Extended data dictionary representing a single item in a commerce transaction, to be included in an event's extended data.
+ 
+ @param itemID The transaction item's ID.
+ @param name The transaction item's name.
+ @param category The transaction item's category.
+ @param price The individual item price.
+ @param quantity The number of units purchased.
+ @param currency Currency for price.
+ */
++ (NSDictionary *)extendedDataCommerceItemWithItemID:(NSString *)itemID
+												name:(NSString *)name
+											category:(NSString *)category
+											   price:(NSNumber *)price
+											quantity:(NSNumber *)quantity
+											currency:(NSString *)currency;
 
 
 ///-------------------------------------
@@ -378,5 +464,12 @@ extern NSString *const ATIntegrationKeyAmazonSNS;
  @param deviceToken The device token expected by AWS SNS.
  */
 - (void)addAmazonSNSIntegrationWithDeviceToken:(NSData *)deviceToken;
+
+/**
+ Adds Parse integration with the given device token.
+ 
+ @param deviceToken The device token expected by Parse.
+ */
+- (void)addParseIntegrationWithDeviceToken:(NSData *)deviceToken;
 
 @end