+++ ApptentiveConnect/source/Engagement/Controllers/ATInteractionFeedbackDialogController.m
@@ -65,7 +65,7 @@ NSString *const ATInteractionFeedbackDialogEventLabelViewMessages = @"view_messa
 		NSString *body = config[@"body"] ?: ATLocalizedString(@"What can we do to ensure that you love our app? We appreciate your constructive feedback.", @"Custom placeholder feedback text when user is unhappy with the application.");
 		messagePanel.promptText = body;
 		
-		BOOL showEmailAddressField = [config[@"ask_for_email"] boolValue] ?: YES;
+		BOOL showEmailAddressField = config[@"ask_for_email"] ? [config[@"ask_for_email"] boolValue] : YES;
 		messagePanel.showEmailAddressField = showEmailAddressField;
 		
 		[messagePanel presentFromViewController:self.viewController animated:YES];
@@ -105,9 +105,19 @@ NSString *const ATInteractionFeedbackDialogEventLabelViewMessages = @"view_messa
 		person = [[[ATPersonInfo alloc] init] autorelease];
 	}
 	if (emailAddress && ![emailAddress isEqualToString:person.emailAddress]) {
-		person.emailAddress = emailAddress;
-		person.needsUpdate = YES;
+		// Do not save empty string as person's email address
+		if (emailAddress.length > 0) {
+			person.emailAddress = emailAddress;
+			person.needsUpdate = YES;
+		}
+		
+		// Deleted email address from form, then submitted.
+		if ([emailAddress isEqualToString:@""] && person.emailAddress) {
+			person.emailAddress = @"";
+			person.needsUpdate = YES;
+		}
 	}
+	
 	[person saveAsCurrentPerson];
 	
 	[[ATBackend sharedBackend] sendTextMessageWithBody:message completion:^(NSString *pendingMessageID) {