+++ ApptentiveConnect/source/Message Center/ATMessageCenterViewController.m
@@ -276,7 +276,7 @@ typedef enum {
 				}
 				textCell.messageBubbleImage.image = chatBubbleImage;
 				
-				textCell.userIcon.image = [ATBackend imageNamed:@"at_mc_user_icon"];
+				textCell.userIcon.image = [ATBackend imageNamed:@"at_mc_user_icon_default"];
 				textCell.usernameLabel.text = ATLocalizedString(@"You", @"User name for text bubbles from users.");
 			} else {
 				textCell = developerCell;
@@ -291,7 +291,7 @@ typedef enum {
 				}
 				textCell.messageBubbleImage.image = chatBubbleImage;
 				
-				textCell.userIcon.image = [ATBackend imageNamed:@"at_mc_user_icon"];
+				textCell.userIcon.image = [ATBackend imageNamed:@"at_mc_user_icon_default"];
 			}
 			[[textCell retain] autorelease];
 			[userCell release], userCell = nil;
@@ -428,7 +428,7 @@ typedef enum {
 			ATFileMessage *fileMessage = (ATFileMessage *)message;
 			[currentCell configureWithFileMessage:fileMessage];
 		}
-		currentCell.userIcon.image = [ATBackend imageNamed:@"at_mc_user_icon"];
+		currentCell.userIcon.image = [ATBackend imageNamed:@"at_mc_user_icon_default"];
 		
 		
 		
@@ -480,9 +480,10 @@ typedef enum {
 		case NSFetchedResultsChangeInsert:
 			[self.tableView insertSections:[NSIndexSet indexSetWithIndex:sectionIndex] withRowAnimation:UITableViewRowAnimationFade];
 			break;
-			
 		case NSFetchedResultsChangeDelete:
 			[self.tableView deleteSections:[NSIndexSet indexSetWithIndex:sectionIndex] withRowAnimation:UITableViewRowAnimationFade];
+			break;
+		default:
 			break;
 	}
 }