+++ ApptentiveConnect/source/Misc/ATJSONSerialization.m
@@ -10,7 +10,18 @@
 
 @implementation ATJSONSerialization
 + (NSData *)dataWithJSONObject:(id)obj options:(ATJSONWritingOptions)opt error:(NSError **)error {
-	return [NSJSONSerialization dataWithJSONObject:obj options:opterr error:error];
+	if ([NSJSONSerialization isValidJSONObject:obj]) {
+		NSData *jsonData = nil;
+		@try {
+			jsonData = [NSJSONSerialization dataWithJSONObject:obj options:opterr error:error];
+		} @catch (NSException *exception) {
+			ATLogError(@"Unable to create JSON data from object: %@ Exception: %@", obj, exception);
+		}
+		return jsonData;
+	} else {
+		ATLogError(@"Attempting to create JSON data from an invalid JSON object.");
+		return nil;
+	}
 }
 
 + (NSString *)stringWithJSONObject:(id)obj options:(ATJSONWritingOptions)opt error:(NSError **)error {